requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
chardet==5.2.0
distro==1.9.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
lxml==5.1.0
openai==1.14.3
pillow==10.2.0
pydantic==2.6.4
pydantic_core==2.16.3
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.1
reportlab==4.1.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.10.0
